pyproject.toml
[project]
name = "ff-iii-toolkit-api"
version = "2.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "fastapi>=0.121.2",
    "fireflyiii-enricher-core",
    "pydantic>=2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "starlette>=0.37.2",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.0.0"
]

[tool.uv.sources]
fireflyiii-enricher-core = { git = "https://github.com/wini83/fireflyiii-enricher-core.git", rev = "main" }


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
[dependency-groups]
dev = [
    "black>=25.11.0",
    "commitizen>=4.10.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "pytest>=9.0.2",
    "ruff>=0.14.5",
]

[tool.deptry]
package_module_name_map = { "pyjwt" = "jwt","python-dotenv" = "dotenv" }

[tool.deptry.per_rule_ignores]
DEP002 = ["python-multipart","uvicorn"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"
